Add `StreamStack#cleanup()`. Which is a function that should be called at the end of the StreamStack instance's life.
I.E. If the StreamStack is only good for a single HTTP request/response, then `cleanup()` should be called after the HTTP response's 'end' event is fired.
For a forever-running StreamStack, like gzip or Icecast metadata, then `cleanup()` may be called when something like a 'close' event happens on the parent stream.